refactor: move `HomePage` and `RootLayout` to FSD-layers
