+++ src/app/layout/ui/RootLayout.tsx
@@ -1,0 +1,34 @@
+import type { Metadata } from 'next';
+import { AppRouterCacheProvider } from '@mui/material-nextjs/v13-appRouter';
+
+import { Inter } from 'next/font/google';
+import { Box } from '@mui/material';
+
+const inter = Inter({ subsets: ['latin'] });
+
+export const metadata: Metadata = {
+  title: 'REST/GraphiQL Client App',
+  description: 'Generated by create next app',
+};
+
+export const RootLayout = ({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) => {
+  return (
+    <Box component="html" height="100%" lang="en">
+      <Box
+        component="body"
+        m={0}
+        height="100%"
+        minHeight="100%"
+        className={inter.className}
+      >
+        <AppRouterCacheProvider>{children}</AppRouterCacheProvider>
+      </Box>
+    </Box>
+  );
+};
+
+export default RootLayout;
