+++ src/pages/HomePage/ui/index.ts
@@ -1,0 +1,3 @@
+import HomePage from './HomePage';
+
+export { HomePage };
